restructure client package
