+++ src/main/java/com/brachy84/mechtech/client/gui/ModuleSlot.java
@@ -1,4 +1,4 @@
-package com.brachy84.mechtech.client;
+package com.brachy84.mechtech.client.gui;
 
 import gregtech.api.gui.widgets.SlotWidget;
 import net.minecraft.item.ItemStack;
